feat: save the current block to local storage
